errors

+++ avr_common/common.h
@@ -4,7 +4,7 @@
 #include <string.h>
 
 typedef struct __attribute__((packed)) amp_value {
-  uint16_t current; 
+  float current; 
   uint16_t timestamp;
 } amp_value;
